refactor common code into SamplingRuleHelper
